requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
black==23.3.0
celery==5.3.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
flake8==6.0.0
iniconfig==2.0.0
kombu==5.3.2
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
